feat: KZG commitments with single strategy for proving
