+++ src/bin/ezkl.rs
@@ -5,26 +5,31 @@ use ezkl::fieldutils::i32_to_felt;
 use ezkl::graph::Model;
 use ezkl::pfsys::ipa::{create_ipa_proof, verify_ipa_proof};
 #[cfg(feature = "evm")]
-use ezkl::pfsys::kzg::{
+use ezkl::pfsys::kzg::aggregation::{
     aggregation::AggregationCircuit, evm_verify, gen_aggregation_evm_verifier,
     gen_application_snark, gen_kzg_proof, gen_pk, gen_srs,
 };
+#[cfg(not(feature = "evm"))]
+use ezkl::pfsys::kzg::single::{create_kzg_proof, verify_kzg_proof};
 use ezkl::pfsys::Proof;
 use ezkl::pfsys::{parse_prover_errors, prepare_circuit_and_public_input, prepare_data};
 #[cfg(feature = "evm")]
 use halo2_proofs::poly::commitment::Params;
+#[cfg(not(feature = "evm"))]
+use halo2_proofs::poly::kzg::{
+    commitment::ParamsKZG, multiopen::VerifierGWC, strategy::SingleStrategy as KZGSingleStrategy,
+};
 use halo2_proofs::{
     dev::MockProver,
     plonk::verify_proof,
     poly::{
         commitment::ParamsProver,
-        ipa::{commitment::ParamsIPA, strategy::SingleStrategy},
+        ipa::{commitment::ParamsIPA, strategy::SingleStrategy as IPASingleStrategy},
         VerificationStrategy,
     },
     transcript::{Blake2bRead, Challenge255, TranscriptReadBuffer},
 };
-#[cfg(feature = "evm")]
-use halo2curves::bn256::G1Affine;
+use halo2curves::bn256::{Bn256, Fr};
 use halo2curves::pasta::vesta;
 use halo2curves::pasta::Fp;
 use log::{error, info, trace};
@@ -81,14 +86,14 @@ pub fn main() {
             model: _,
             pfsys,
         } => {
+            // A direct proof
             let args = Cli::parse();
             let data = prepare_data(data);
-            let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
-            info!("full proof with {}", pfsys);
-
             match pfsys {
                 ProofSystem::IPA => {
-                    // A direct proof
+                    let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
+                    info!("full proof with {}", pfsys);
+
                     let params: ParamsIPA<vesta::Affine> = ParamsIPA::new(args.logrows);
                     trace!("params computed");
 
@@ -103,7 +108,7 @@ pub fn main() {
                     let pi_for_real_prover: &[&[&[Fp]]] = &[&pi_inner];
 
                     let now = Instant::now();
-                    let strategy = SingleStrategy::new(&params);
+                    let strategy = IPASingleStrategy::new(&params);
                     let mut transcript = Blake2bRead::<_, _, Challenge255<_>>::init(&proof[..]);
                     assert!(verify_proof(
                         &params,
@@ -116,11 +121,40 @@ pub fn main() {
                     info!("verify took {}", now.elapsed().as_secs());
                 }
                 #[cfg(not(feature = "evm"))]
-                ProofSystem::KZG => todo!(),
+                ProofSystem::KZG => {
+                    // A direct proof
+                    let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
+                    let params: ParamsKZG<Bn256> = ParamsKZG::new(args.logrows);
+                    trace!("params computed");
+
+                    let (pk, proof, _dims) =
+                        create_kzg_proof(circuit, public_inputs.clone(), &params);
+
+                    let pi_inner: Vec<Vec<Fr>> = public_inputs
+                        .iter()
+                        .map(|i| i.iter().map(|e| i32_to_felt::<Fr>(*e)).collect::<Vec<Fr>>())
+                        .collect::<Vec<Vec<Fr>>>();
+                    let pi_inner = pi_inner.iter().map(|e| e.deref()).collect::<Vec<&[Fr]>>();
+                    let pi_for_real_prover: &[&[&[Fr]]] = &[&pi_inner];
+
+                    let now = Instant::now();
+                    let strategy = KZGSingleStrategy::new(&params);
+                    let mut transcript = Blake2bRead::<_, _, Challenge255<_>>::init(&proof[..]);
+                    assert!(verify_proof::<_, VerifierGWC<_>, _, _, _>(
+                        &params,
+                        pk.get_vk(),
+                        strategy,
+                        pi_for_real_prover,
+                        &mut transcript
+                    )
+                    .is_ok());
+                    info!("verify took {}", now.elapsed().as_secs());
+                }
                 #[cfg(feature = "evm")]
                 ProofSystem::KZG => {
                     // We will need aggregator k > application k > bits
                     //		    let application_logrows = args.logrows; //bits + 1;
+                    let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
                     let aggregation_logrows = args.logrows + 6;
 
                     let params = gen_srs(aggregation_logrows);
@@ -164,59 +198,117 @@ pub fn main() {
         } => {
             let args = Cli::parse();
             let data = prepare_data(data);
-            let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
-            info!("proof with {}", pfsys);
-            let params: ParamsIPA<vesta::Affine> = ParamsIPA::new(args.logrows);
-            trace!("params computed");
+            match pfsys {
+                ProofSystem::IPA => {
+                    let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
+                    info!("proof with {}", pfsys);
+                    let params: ParamsIPA<vesta::Affine> = ParamsIPA::new(args.logrows);
+                    trace!("params computed");
 
-            let (_pk, proof, _input_dims) =
-                create_ipa_proof(circuit.clone(), public_inputs.clone(), &params);
+                    let (_pk, proof, _input_dims) =
+                        create_ipa_proof(circuit.clone(), public_inputs.clone(), &params);
 
-            let pi: Vec<_> = public_inputs
-                .into_iter()
-                .map(|i| i.into_iter().collect())
-                .collect();
+                    let pi: Vec<_> = public_inputs
+                        .into_iter()
+                        .map(|i| i.into_iter().collect())
+                        .collect();
 
-            let checkable_pf = Proof {
-                input_shapes: circuit.inputs.iter().map(|i| i.dims().to_vec()).collect(),
-                public_inputs: pi,
-                proof,
-            };
+                    let checkable_pf = Proof {
+                        input_shapes: circuit.inputs.iter().map(|i| i.dims().to_vec()).collect(),
+                        public_inputs: pi,
+                        proof,
+                    };
 
-            let serialized = match serde_json::to_string(&checkable_pf) {
-                Ok(s) => s,
-                Err(e) => {
-                    abort!("failed to convert proof json to string {:?}", e);
+                    let serialized = match serde_json::to_string(&checkable_pf) {
+                        Ok(s) => s,
+                        Err(e) => {
+                            abort!("failed to convert proof json to string {:?}", e);
+                        }
+                    };
+
+                    let mut file = std::fs::File::create(output).expect("create failed");
+                    file.write_all(serialized.as_bytes()).expect("write failed");
                 }
-            };
+                ProofSystem::KZG => {
+                    let (circuit, public_inputs) = prepare_circuit_and_public_input(&data);
+                    info!("proof with {}", pfsys);
+                    let params: ParamsKZG<Bn256> = ParamsKZG::new(args.logrows);
+                    trace!("params computed");
 
-            let mut file = std::fs::File::create(output).expect("create failed");
-            file.write_all(serialized.as_bytes()).expect("write failed");
+                    let (_pk, proof, _input_dims) =
+                        create_kzg_proof(circuit.clone(), public_inputs.clone(), &params);
+
+                    let pi: Vec<_> = public_inputs
+                        .into_iter()
+                        .map(|i| i.into_iter().collect())
+                        .collect();
+
+                    let checkable_pf = Proof {
+                        input_shapes: circuit.inputs.iter().map(|i| i.dims().to_vec()).collect(),
+                        public_inputs: pi,
+                        proof,
+                    };
+
+                    let serialized = match serde_json::to_string(&checkable_pf) {
+                        Ok(s) => s,
+                        Err(e) => {
+                            abort!("failed to convert proof json to string {:?}", e);
+                        }
+                    };
+
+                    let mut file = std::fs::File::create(output).expect("create failed");
+                    file.write_all(serialized.as_bytes()).expect("write failed");
+                }
+            }
         }
         Commands::Verify {
             model: _,
             proof,
-            pfsys: _,
-        } => {
-            let mut file = match File::open(proof) {
-                Ok(f) => f,
-                Err(e) => {
-                    abort!("failed to open proof file {:?}", e);
-                }
-            };
-            let mut data = String::new();
-            match file.read_to_string(&mut data) {
-                Ok(_) => {}
-                Err(e) => {
-                    abort!("failed to read file {:?}", e);
-                }
-            };
-            let proof: Proof = serde_json::from_str(&data).expect("JSON was not well-formatted");
+            pfsys,
+        } => match pfsys {
+            ProofSystem::IPA => {
+                let mut file = match File::open(proof) {
+                    Ok(f) => f,
+                    Err(e) => {
+                        abort!("failed to open proof file {:?}", e);
+                    }
+                };
+                let mut data = String::new();
+                match file.read_to_string(&mut data) {
+                    Ok(_) => {}
+                    Err(e) => {
+                        abort!("failed to read file {:?}", e);
+                    }
+                };
+                let proof: Proof =
+                    serde_json::from_str(&data).expect("JSON was not well-formatted");
 
-            let result = verify_ipa_proof(proof);
-            info!("verified: {}", result);
-            assert!(result);
-        }
+                let result = verify_ipa_proof(proof);
+                info!("verified: {}", result);
+                assert!(result);
+            }
+            ProofSystem::KZG => {
+                let mut file = match File::open(proof) {
+                    Ok(f) => f,
+                    Err(e) => {
+                        abort!("failed to open proof file {:?}", e);
+                    }
+                };
+                let mut data = String::new();
+                match file.read_to_string(&mut data) {
+                    Ok(_) => {}
+                    Err(e) => {
+                        abort!("failed to read file {:?}", e);
+                    }
+                };
+                let proof: Proof =
+                    serde_json::from_str(&data).expect("JSON was not well-formatted");
+
+                let result = verify_kzg_proof(proof);
+                info!("verified: {}", result);
+                assert!(result);
+            }
+        },
     }
 }
